requirements.txt
google-generativeai>=0.8.0
python-dotenv
schedule>=1.2.0
requests>=2.31.0
